fix: controller moved and checkstyle update
