--- src/main/java/ru/epa/epabackend/exception/NotFoundException.java
@@ -1,16 +1,0 @@
-package ru.epa.epabackend.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-/**
- * Класс NotFoundException содержит обработку исключения.
- *
- * @author Артем Масалкин
- */
-@ResponseStatus(value = HttpStatus.NOT_FOUND)
-public class NotFoundException extends RuntimeException {
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
