+++ src/main/java/ru/epa/epabackend/service/impl/TechnologyServiceImpl.java
@@ -3,8 +3,8 @@ package ru.epa.epabackend.service.impl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.epa.epabackend.dto.TechnologyDto;
-import ru.epa.epabackend.exception.NotFoundException;
+import ru.epa.epabackend.dto.technology.TechnologyDto;
+import ru.epa.epabackend.exception.exceptions.NotFoundException;
 import ru.epa.epabackend.mapper.TechnologyMapper;
 import ru.epa.epabackend.model.Technology;
 import ru.epa.epabackend.repository.TechnologyRepository;
